requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
bootstrap4==0.1.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
Django==2.2.9
django-bootstrap4==3.0.1
django-chartjs==2.2.1
django-filter==2.4.0
djangorestframework==3.12.4
geoip2==4.3.0
idna==3.2
Markdown==3.3.4
maxminddb==2.1.0
multidict==5.1.0
psutil==5.8.0
pyenv-win==2.64.11
pytz==2021.1
requests==2.26.0
soupsieve==2.2.1
sqlparse==0.4.1
style==1.1.0
typing-extensions==3.10.0.2
update==0.0.1
urllib3==1.26.6
yarl==1.6.3
